Refactor: split providers, decompose retrieval, extract deduplicator, add circuit breaker
Architecture cleanup bringing production-grade improvements:

- WS1: Split core/providers.py (647 lines) into core/providers/ package
  (base.py, _retry.py, litellm.py, openrouter.py) with jittered backoff
  and astream() on LLMInterface

- WS2: Add circuit breaker (core/circuit_breaker.py) for cascading-failure
  protection; integrate into connection.py with jittered exponential backoff

- WS3: Decompose multi_path.py (569→239 lines) into 4 focused modules:
  entity_discovery.py, relationship_expansion.py, chunk_retrieval.py,
  result_assembly.py — each independently testable

- WS4: Extract EntityDeduplicator from api/main.py (576→403 lines) into
  storage/deduplicator.py with label-aware grouping preserved

All 460 tests pass. Lite benchmark: 8.50/10 (85.0%).
No logic, Cypher, store interface, or public API changes.
