+++ graphrag_sdk/src/graphrag_sdk/core/providers/__init__.py
@@ -1,0 +1,20 @@
+# GraphRAG SDK 2.0 — Provider package
+# Re-exports all provider ABCs and built-in implementations.
+
+from graphrag_sdk.core.providers.base import (
+    Embedder,
+    LLMBatchItem,
+    LLMInterface,
+)
+from graphrag_sdk.core.providers.litellm import LiteLLM, LiteLLMEmbedder
+from graphrag_sdk.core.providers.openrouter import OpenRouterEmbedder, OpenRouterLLM
+
+__all__ = [
+    "Embedder",
+    "LLMBatchItem",
+    "LLMInterface",
+    "LiteLLM",
+    "LiteLLMEmbedder",
+    "OpenRouterEmbedder",
+    "OpenRouterLLM",
+]
