+++ graphrag_sdk/src/graphrag_sdk/storage/__init__.py
@@ -1,7 +1,8 @@
 # GraphRAG SDK 2.0 — Storage
 # Data access layer: graph store and vector store.
 
+from graphrag_sdk.storage.deduplicator import EntityDeduplicator
 from graphrag_sdk.storage.graph_store import GraphStore
 from graphrag_sdk.storage.vector_store import VectorStore
 
-__all__ = ["GraphStore", "VectorStore"]
+__all__ = ["EntityDeduplicator", "GraphStore", "VectorStore"]
